refactor(web): extract bulk-page constants and CSV helpers; add parseCSV tests
Move the option lists (tone/provider/strategy/export) to app/bulk/constants.ts
and the pure parseCSV/createDraftItem helpers to app/bulk/csv.ts.
BulkGenerationPageClient.tsx drops from 1164 to ~1074 lines. Add 7 vitest tests
covering parseCSV edge cases and createDraftItem defaults. Behavior-preserving:
tsc, eslint, and all 192 frontend tests pass.
